+++ CHANGELOG.md
@@ -5,7 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Add
+- Ability to print services to stdout one service per line
+- Ability to print endpoints to stdout one endpoint per line
+- Ability to print to a yaml file
 
-
+### Change
 
 [Unreleased]: https://github.com/mattyjones/oathkeeper/compare/0.0.1...HEAD